Parse the code to modules and import them to index.js
