+++ src/TaskListClass.js
@@ -1,3 +1,6 @@
+import renderList from './RenderList.js';
+import taskObject from './TaskObject.js';
+
 class TaskList {
   constructor() {
     this.tasks = JSON.parse(localStorage.getItem('To-Do List')) || [];
@@ -9,7 +12,7 @@ class TaskList {
       event.preventDefault();
       const description = document.getElementById('add-task').value;
       if (description !== '') {
-        const aTask = new taskObject (description);
+        const aTask = new taskObject (description, this.tasks.length);
         this.addTask(aTask);
         renderList(this.tasks);
         document.getElementById('add-task').value = '';
@@ -28,49 +31,4 @@ class TaskList {
   };
 };
 
-const taskList = new TaskList();
-taskList.init();
-
-class taskObject {
-  constructor(description) {
-    this.description = description;
-    this.completed = false;
-    this.index = taskList.tasks.length + 1;
-  }
-};
-
-const listName = "Today's To Do";
-const label = document.querySelector('label');
-label.innerHTML = listName;
-
-const renderList = (arr) => {
-  const ul = document.querySelector('ul');
-  ul.innerHTML = '';
-
-  const sortedArr = [...arr];
-  sortedArr.sort((a, b) => a.index - b.index);
-
-  for (let i = 0; i < sortedArr.length; i += 1) {
-    const li = document.createElement('li');
-    const label = document.createElement('label');
-    const input = document.createElement('input');
-    input.setAttribute('type', 'checkbox');
-
-    const checkmark = document.createElement('span');
-    checkmark.classList.add('checkmark');
-    
-    const moveBtn = document.createElement('button');
-
-    const iconInBtn = document.createElement('img');
-    iconInBtn.src = './drag-handle-minor-svgrepo-com.svg';
-    moveBtn.appendChild(iconInBtn);
-
-    label.appendChild(input);
-    label.appendChild(checkmark);
-    label.innerHTML += `${sortedArr[i].description}`;
-    label.appendChild(moveBtn);
-
-    li.appendChild(label);
-    ul.appendChild(li);
-  }
-};
+export default TaskList;
