+++ src/TaskObject.js
@@ -1,0 +1,9 @@
+class taskObject {
+  constructor(description, index) {
+    this.description = description;
+    this.completed = false;
+    this.index = index + 1;
+  }
+};
+
+export default taskObject;
